adjusted playingGameState
